requirements.txt
requests
pandas
python-dateutil
tqdm
